La classe Calcul contient trois fonctions : moyenne, ecarttype, mediane

+++ Tris.java
@@ -4,9 +4,9 @@ public class Tris{
 
   }
 
-  public void method1(int[] numbers){
+  public void method1(double[] numbers){
     for (int i=1;i<numbers.length ;i++ ) {
-      int save=numbers[i];
+      double save=numbers[i];
       int j=i;
       while (j>0 && numbers[j-1]>save) {
         numbers[j]=numbers[j-1];
@@ -17,7 +17,7 @@ public class Tris{
     return;
   }
 
-  public void method2(int[] numbers){
+  public void method2(double[] numbers){
     for (int i=0;i<numbers.length-1 ;i++ ) {
       int min=i;
       for (int j=i+1;j<numbers.length; j++ ) {
@@ -26,18 +26,18 @@ public class Tris{
         }
       }
       if (min!=i) {
-        int save=numbers[i];
+        double save=numbers[i];
         numbers[i]=numbers[min];
         numbers[min]=save;
       }
     }
   }
 
-  public void method3(int[] numbers){
+  public void method3(double[] numbers){
     for (int i=numbers.length; i>1; i--) {
       for (int j=0;j<i-1 ;j++ ) {
         if (numbers[j+1]<numbers[j]) {
-          int save=numbers[j];
+          double save=numbers[j];
           numbers[j]=numbers[j+1];
           numbers[j+1]=save;
         }
